Added Keypad/Gave Coord gridPosition()
Added a new structure to help manage large groups of buttons. Menu now inherits from this, and this new structure is Menu minus Pages. Wrote some small but effective methods in Coord to make 1D to 2D grid placement conversions a bit less hands-on. Needed good documentation though, as the methods are a bit weird.